# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LASU Fleet Manager API",
        "description": "API documentation for the LASU fleet management system.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auditor/logs": {
            "get": {
                "tags": [
                    "Auditor"
                ],
                "summary": "Get audit logs with optional filters",
                "operationId": "7411baad940a0eb3222de21adbd62f0e",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type of log",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "fuel",
                                "maintenance",
                                "condition"
                            ]
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Start date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "End date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of audit logs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "logId": {
                                                "type": "integer",
                                                "example": 9001
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "fuel"
                                            },
                                            "requestId": {
                                                "type": "integer",
                                                "example": 101
                                            },
                                            "changedBy": {
                                                "type": "string",
                                                "example": "manager"
                                            },
                                            "timestamp": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-07-20T10:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User login using plate number and password",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plateNumber",
                                    "password"
                                ],
                                "properties": {
                                    "plateNumber": {
                                        "type": "string",
                                        "example": "LASU-123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "jwt-token"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 42
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "owner"
                                                },
                                                "plateNumber": {
                                                    "type": "string",
                                                    "example": "LASU-123"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
        "/api/manager/fuel-requests/{requestId}": {
            "patch": {
                "tags": [
                    "Manager"
                ],
                "summary": "Approve or reject a fuel request",
                "operationId": "2b756b6721493b2b85d2191aa43083df",
                "parameters": [
                    {
                        "name": "requestId",
                        "in": "path",
                        "description": "Fuel request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "action"
                                ],
                                "properties": {
                                    "action": {
                                        "type": "string",
                                        "enum": [
                                            "approve",
                                            "reject"
                                        ],
                                        "example": "approve"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 101
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "approved"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/manager/maintenance-requests/{requestId}": {
            "patch": {
                "tags": [
                    "Manager"
                ],
                "summary": "Approve or reject a maintenance request",
                "operationId": "f61211d3acd5ac5e9f7369d1aef25742",
                "parameters": [
                    {
                        "name": "requestId",
                        "in": "path",
                        "description": "Maintenance request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "action"
                                ],
                                "properties": {
                                    "action": {
                                        "type": "string",
                                        "enum": [
                                            "approve",
                                            "reject"
                                        ],
                                        "example": "reject"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 202
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "rejected"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}